Use VideoWidth instead of VideoHeight for IS_SD, IS_HD, IS_4K etc...
For example 4K letterbox did not had the correct indication as it might
have less then 2100 lines. During a discussion on the forum we concluce
it is better to take the VideoWidth instead for this decisian.

https://forums.openpli.org/topic/82219-scaler-and-resolution/page-2
(cherry picked from commit 3f04a07a9618a8aad07c427c1dcb8d5501405f89)