[ChannelSelection] Optimized string for translation
Not sure why a 4 variable string (%d.%d%s%s) was needed to display the orbital position... This was misleading (couple of translations had errors) and was producing the following warning as well:

'msgid' format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

The new string is using already existing ones for the orbital position ("%.1f° W" and "%.1f° E") and is much simpler (only one variable is used for the orbital posiiton in the final string).

All existing translations are updated as needed.

(cherry picked from commit d69e21814636f16cc4b12809b3b2928943ac3513)