[InputBox] Reformat text string
Like it was split before, it was not clear to translators what was the meaning of the message. It was unclear what was the "%s" that "You have to wait!" for and translations were erroneous. So, reformatted the message to consist of one string, not 3 different ones and force the translations to be done again, properly this time.

po files are NOT updated on purpose!

(cherry picked from commit 2e3e8abc29632fb06ecc8a7804a2cfc895f90603)