Patch fastscan (#3118)
* [FastScan] Use named arguments for string

This fixes the following warning when updating the po files:

'msgid' format string with unnamed arguments cannot be properly localized:
The translator cannot reorder the arguments.
Please consider using a format string with named arguments,
and a mapping instead of a tuple for the arguments.

All existing translations are updated with the new string formatting.

* [FastScan] Use better wording for text message

The old string will be removed from the translations as it's not used elsewhere.
The new string is already present in the translations, so no update is needed to the po files.

(cherry picked from commit 544c3ff211dc86c6a87df23c8bafd8460b0676b4)