Use boolean checks according to PEP8
One “bad” way: if variable == True:
Another “bad” way: if variable is True:
And the good way, recommended even in the Programming Recommendations of PEP8: if variable:

The if variable and if not variable versions are faster to execute and faster to read.

Also fix few more SyntaxWarning: "is" with a literal. Did you mean "=="?

(cherry picked from commit 57305a3a183f96efbd54742bf10d53bf24322787)