[Navigation] Streaming: Retry play service
If a stream is active which blocks a tuner on the same box,
changing the channel might not work. The stream client is stopped
before zapping, but the streamserver still blocks the tuner when e2
tries to allocate the tuner.
Use an eTimer to give the streamserver the time to release the tuner
and then try again.

(cherry picked from commit af508c35038f54092a7c89e8e42088a506da4480)