Ensure and unrequired variable has a default value in ConfigIPUnreq
The function getHTML is usefull here to get the IP-address from a
ConfigIP in a string instead of a list of integer. But the function
created for it needs a parameter. But this parameter is not used at all
in the specific function. So I give it a default so it becomes not
mandatory. But I need to keep it as I'm not aware where else it might be
used when this function is called (especially in plugins etc)

(cherry picked from commit d318bc45be971d003594a47b8fb61839ec9164a5)