When streamrelay server is local host then modify the service reference
It seems this is needed for the current oscam binaries but it is not
clear why. It also makes it less transparant. It worked when it was used
on a fallback clinet where I developed the feature on, but not on the
box where the oscam binary is really running on. I think we still need
to investigate why we need a fake service ref here and try to find a way
to resolve it that it is not required. To avoid an extra config I used
the local host for the detecting when it is needed localy. So on
fallback clients you automatically get better data in the OSD.

Thanks Huevos for the initial report and sorry that testing and investigating
took a while.

Now I still hope the community helps to resolve that this is required so
we can revert it. Let's hope as for the OSD you will get more valuable
values.

(cherry picked from commit b32d29c903f5f964e1f08d57a450d28654edb01d)