Fix saving the set range in ProgressBar
When setting the progress bar range, it is set in eSlider but not stored in instance variables.
It does not allow to get the correct range using the ProgressBar method getRange or property range.
Therefore, when setting the range, also stores it in instance variables.

(cherry picked from commit 3b4c82e5a27da555c5145de375f7bb09839df4ab)