Streamline checksum check regarding the BoxInfo
This key should always be there. When the "checksum" key's value is None
then there is no /usr/lib/enigma.info file available or the file does
not contain a checksum (in the last) row. When it is False the file is
manupulated without adapting the checksum and when it is True the file
is completely as it should be. This key in the BoxInfo class can always
be used to give a popup as warning later on.

(cherry picked from commit 35c2d8eab5debcd5a5f72d2af1fcb0db9670c5fd)