Fix Last-Modified date in softwareupdate
From the website the timestamps came in English and when the box is in a
different local language this will go wrong. Now I set temporary the
locale.LC_TIME back to English before converting the datestring from the
website and then afterwards set the language back to it was.

We still may consider to get this improved.

(cherry picked from commit 561745d5bf92b07af84f89b5362db52ce9b087cd)