Ensure SystemInfo dict is the same as BoxInfo.boxInfoMutable dict
In the suggested code for the BoxInfo SystemInfo was a duplicate so in
fact the mutable items were stored twice. Now they are stored one time.
In addition when we later do not have SystemInfo in enigma2 and it's
plugins we can remove this line (and likely the imports for multiboot
can be replaced on top)

(cherry picked from commit f7db9ecf47a63edb8081d567bb448ae9a8b296e1)