Revert "When setItem is used ensure it is always available via SystemInfo"
This reverts commit 713f43e3abbcbc43f86f3e6acac0928406d716f8.

I still think this is not a good idea.

At this moment only the new items coming from enigma.info are not
available in SystemInfo as key and we defined now the data from
enigma.info as immutable. These values are usually new values that
always can be obtained via BoxInfo.getItem.So when no code is generated
it can always use this newer function to obtain the data and no real
need to obtain it as a SystemInfo key is needed. As it is now at least
all current SystemInfo keys are now also available via BoxInfo where
they are mutable and exactly the same. It does not care when you try to
manipulate them via SystemInfo or BoxInfo.

(cherry picked from commit 3dede665ad8425c6b5f0bb31b7c0f0f1430c84f5)