ImportChannels: Do not break the import when a file cannot be fatched
Currently in the hanssat settings there is a file that is not available 'userbouquet.stream_solomonseilanden__sb_.radio' it should be but in a bouquet the string is different. Currently it breaks the total import. This change does avoid that so at least all the rest is read.

(cherry picked from commit 9abf89a1ec339271ee5da63fdbfa27a6bbb58f46)