Add setting change for secondinfobar simple and infobar
Previously it could only be toggled instantly, but users might better
use it when it is also visible in the config setting. Also arrange that
with conditional you can show setups when specific screens are available
in a skin or not show them when they are not available.

(cherry picked from commit d0fe68db7d79158e851a5056cf5ba479aae1a2f2)