With ImportChannels when there is no route to host try 3 times
It could be that on a remote box when e.g. it is connected to WLAN and
that box is in standby that the WLAN network is in a kind of sleep mode
and with the first attempt the WLAN needs to wake-up somehow and the box
is not reachable. Therefore I try it 3 times when there is no route to a
host give time for that remote box to wakeup and get the WLAN online so
there is some time for that box to respond.

(cherry picked from commit 2aba91ea47bc349cee45b8b482bf0c5541f7e9c9)