Resolve removal of debuglog files in troubleshoot screen
The removal of log files was based on the 'cat filename.log' string
where we removed the cat. But debug files do 'tail -n 2500
filename.debuglog". When we remove the cat it will become ' -n 2500
filename.debuglog" as filename as that does not work. So I hope it is
sufficiant to resolve it by stripping with rstrip(" ', 1)[1]. For all
log files this should be sufficiant.

(cherry picked from commit 1e73ac267484f76507df4e8696f4fa58b8cba774)