Resolve status indicator for Fallback Timers
Previously the green button was not updated on all the screens where the
fallback timers were shown. Now the callback is replaced so it is called
after the timers are changed so the checks for the green buttons etc. do
work

(cherry picked from commit 2ddbcfce6b4c5b6ff72224753f5ce4b4594e3857)