Remove option installing picons to an alternative location
You cannot use -add-dest with opkg update because when new packages are
included then they will be installed on an inappropriate location.

For more information see on the forum here:
https://forums.openpli.org/topic/102790-vu-ultimo-4k-extensionopenwebif-nonetype-object-is-not-subscriptable/#entry1694446

Over the year we get manu reports that users try to install packages and
then it was not possible for them while they are within the feed.

For now I removed the complete option. Maybe later we could re-add this
feature, but then in a different way.

(cherry picked from commit 77a7ebe151f40a637073d41e3ad153ec403b4344)